+++ guillotina/db/storages/pg.py
@@ -402,7 +402,9 @@ class PostgresqlStorage(BaseStorage):
         'CREATE SEQUENCE IF NOT EXISTS tid_sequence;'
     ]
 
-    _unique_constraint = 'ALTER TABLE objects ADD CONSTRAINT objects_parent_id_id_key UNIQUE (parent_id, id)'
+    # _unique_constraint = '''ALTER TABLE objects
+    #                         ADD CONSTRAINT objects_parent_id_id_key
+    #                         UNIQUE (parent_id, id)'''
 
     def __init__(self, dsn=None, partition=None, read_only=False, name=None,
                  pool_size=13, transaction_strategy='resolve_readcommitted',
@@ -491,13 +493,13 @@ class PostgresqlStorage(BaseStorage):
             await self.initialize_tid_statements()
             await self._read_conn.execute(CREATE_TRASH)
 
-        try:
-            await self._read_conn.execute(self._unique_constraint)
-        except asyncpg.exceptions.DuplicateTableError:
-            pass
-        except asyncpg.exceptions.InternalServerError as ex:
-            if 'duplicate constraint name' not in ex.message:
-                raise
+        # try:
+        #     await self._read_conn.execute(self._unique_constraint)
+        # except asyncpg.exceptions.DuplicateTableError:
+        #     pass
+        # except asyncpg.exceptions.InternalServerError as ex:
+        #     if 'duplicate constraint name' not in ex.message:
+        #         raise
 
         # migrate to larger VARCHAR size...
         result = await self._read_conn.fetch("""